Rename view from home to transcriptions
