+++ src/views/TranscriptionsView.vue
